% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dino_list.R
\docType{data}
\name{dino_list}
\alias{dino_list}
\title{dino list}
\format{An object of class \code{data.frame} with 1767 rows and 4 columns.}
\source{
\url{https://en.wikipedia.org/wiki/List_of_dinosaur_genera}
}
\usage{
dino_list
}
\description{
dino list
}
\keyword{datasets}
